Cross-build the plugin for sbt 1.x and sbt 2.x
sbt 2.0.0 runs on Scala 3, which has a different macro system than the
Scala 2 `macro`/blackbox.Context one powering the `module` macro. To
publish from a single codebase for both sbt majors:

- Extract `module` into a per-axis `ModuleMacroCompat` trait that
  `autoImport` mixes in. The shared sources keep everything else; only
  the macro entry point differs per Scala version.
  - scala-2.12: the existing blackbox `Macros` bundle + `KeyMacro`.
  - scala-3: a `scala.quoted` macro modelled on sbt 2's own
    `sbt.std.KeyMacro` (same 'copied from sbt' lineage as before).
- Add the Scala 3 axis: `crossScalaVersions` includes 3.8.4 (sbt
  2.0.0's Scala) and `pluginCrossBuild / sbtVersion` maps the 2.12 axis
  to sbt 1.12.12 (the JDK-25-capable floor that scripted needs) and the
  Scala 3 axis to 2.0.0.
- Mark `moduleMetadata` `@transient`: sbt 2 tries to cache task results
  and the result is derived from the whole build `state`, which is not
  cacheable. `@transient` opts out (valid on both axes, no extra dep).
- The meta-build (Scala 2.12) now also reads the scala-2.12 source dir
  so `build.sbt`'s own `module` call keeps resolving.
- Cross-build the ci-test alias steps with `+`.
