+++ modules/sbt-modules/src/main/scala-2.12/com/alejandrohdezma/sbt/modules/KeyMacro.scala
