+++ project/dependencies.sbt
@@ -1,2 +1,4 @@
-Compile / unmanagedSourceDirectories +=
-  (ThisBuild / baseDirectory).value.getParentFile / "modules" / "sbt-modules" / "src" / "main" / "scala"
+Compile / unmanagedSourceDirectories ++= {
+  val main = (ThisBuild / baseDirectory).value.getParentFile / "modules" / "sbt-modules" / "src" / "main"
+  Seq(main / "scala", main / "scala-2.12")
+}
